+++ tests/run_examples.sh
@@ -93,6 +93,7 @@ examples=(
     record_keyword_keys_test.bas
     datetime_test.gb
     datetime_exact_comparison_test.bas
+    datetime_arithmetic_test.bas
     datetime_lens_test.bas
     datetime_lens_operator_test.bas
     datetime_modifier_test.bas